Moved checkIfFileMatchesLayer to its own place
