+++ adagucserverEC/adagucserver.cpp
@@ -37,6 +37,7 @@
 #include "utils/ConfigurationUtils.h"
 #include "CDBFactory.h"
 #include "CConvertGeoJSON.h"
+#include "utils/serverutils.h"
 
 DEF_ERRORMAIN();
 
@@ -81,33 +82,6 @@ void serverWarningFunction(const char *msg) {
 }
 
 void serverLogFunctionCMDLine(const char *msg) { printf("%s", msg); }
-
-bool checkIfFileMatchesLayer(CT::string layerPathToScan, CServerConfig::XMLE_Layer *layer) {
-  // Get the directory of the file to scan:
-  CT::string directoryOfFileToScan = layerPathToScan = CDirReader::makeCleanPath(layerPathToScan);
-  directoryOfFileToScan.substringSelf(0, directoryOfFileToScan.length() - directoryOfFileToScan.basename().length());
-  directoryOfFileToScan = CDirReader::makeCleanPath(directoryOfFileToScan) + "/";
-
-  if (layer->attr.type.empty() || layer->attr.type.equals("database")) {
-    if (layer->FilePath.size() > 0) {
-      CT::string filePath = CDirReader::makeCleanPath(layer->FilePath[0]->value);
-      // Directories need to end with a /
-      CT::string filePathWithTrailingSlash = filePath + "/";
-      CT::string filter = layer->FilePath[0]->attr.filter;
-      // CDBDebug(" %s %s", directoryOfFileToScan.c_str(), directoryOfFileToScan.c_str());
-      // When the FilePath in the Layer configuration is exactly the same as the file to scan, give a Match
-      if (layerPathToScan.equals(filePath)) {
-        return true;
-        // When the directory of the file to scan matches the FilePath and the filter matches, give a Match
-      } else if (directoryOfFileToScan.startsWith(filePathWithTrailingSlash)) {
-        if (CDirReader::testRegEx(layerPathToScan.basename(), filter.c_str()) == 1) {
-          return true;
-        }
-      }
-    }
-  }
-  return false;
-}
 
 /**
  * @param layerPathToScan: the provided file to scan
