+++ adagucserverEC/utils/serverutils.h
@@ -1,0 +1,9 @@
+#ifndef SERVERUTILS_H
+#define SERVERUTILS_H
+
+#include "CTString.h"
+#include "CServerConfig_CPPXSD.h"
+
+bool checkIfFileMatchesLayer(CT::string layerPathToScan, CServerConfig::XMLE_Layer *layer);
+
+#endif
